hooks 10

+++ backend/Application/Tasks/Commands/ClearCompletedTasks.cs
@@ -3,12 +3,14 @@ using Application.Common.Security;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Application.Common.Security.Attributes;
+using Application.Common.Exceptions;
 
 namespace Application.Tasks.Commands.ClearCompletedTasks
 {
     [Authorize]
     public class ClearCompletedTasksCommand : IRequest
     {
+        public int TaskListId { get; set; }
     }
 
     public class ClearCompletedTasksCommandHandler : IRequestHandler<ClearCompletedTasksCommand>
@@ -28,15 +30,27 @@ namespace Application.Tasks.Commands.ClearCompletedTasks
         {
             var currentUserId = int.Parse(_currentUserService.UserId!);
 
-            var completedTasks = await _context.Tasks
-.Include(t => t.TaskList)
-.ThenInclude(tl => tl.UserTaskLists)
-.Where(t => t.IsCompleted && t.TaskList.UserTaskLists.Any(utl => utl.UserId == currentUserId))
-.ToListAsync(cancellationToken);
+            var taskList = await _context.TaskLists
+                .Include(tl => tl.UserTaskLists)
+                .FirstOrDefaultAsync(tl => tl.Id == request.TaskListId, cancellationToken);
 
+            if (taskList == null)
+            {
+                throw new NotFoundException(nameof(Domain.Entities.TaskList), request.TaskListId);
+            }
+
+            if (!taskList.UserTaskLists.Any(utl => utl.UserId == currentUserId && !utl.IsDeleted))
+            {
+                throw new ForbiddenAccessException();
+            }
+
+            var completedTasks = await _context.Tasks
+                .Where(t => t.TaskListId == request.TaskListId && t.IsCompleted)
+                .ToListAsync(cancellationToken);
 
             _context.Tasks.RemoveRange(completedTasks);
             await _context.SaveChangesAsync(cancellationToken);
+
             foreach (var task in completedTasks)
             {
                 await _notificationService.SendTaskDeletedNotification(task.TaskListId, task.Id);
